requirements.txt
flask
flask-login
flask-sqlalchemy
reportlab

///pip install -r requirements.txt
